restructure incremental test to permit more tests
